+++ GeneralsMD/Code/GameEngine/Include/GameClient/GadgetPushButton.h
@@ -83,6 +83,8 @@ void GadgetButtonDrawInverseClock( GameWindow *g, Int percent, Color color ); //
 void GadgetButtonDrawOverlayImage( GameWindow *g, const Image *image );
 void GadgetButtonSetBorder( GameWindow *g, Color color, Bool drawBorder = TRUE );
 void GadgetButtonSetCount( GameWindow *g, Int count );
+void GadgetButtonSetSeconds( GameWindow *g, Int seconds );
+void GadgetButtonSetBar( GameWindow *g, Int percent, Color color );
 void GadgetButtonSetData(GameWindow *g, void *data);
 void *GadgetButtonGetData(GameWindow *g);
 void GadgetButtonSetAltSound( GameWindow *g, AsciiString altSound );