+++ GeneralsMD/Code/GameEngine/Source/GameClient/MessageStream/GUICommandTranslator.cpp
@@ -287,6 +287,7 @@ static CommandStatus doAttackMoveCommand( const CommandButton *command, const IC
 	// send the message to set the rally point
 	GameMessage *msg = TheMessageStream->appendMessage( GameMessage::MSG_DO_ATTACKMOVETO );
 	msg->appendLocationArgument( world );
+	msg->appendBooleanArgument( TheInGameUI->isInForceAttackMode() );	// ctrl: arrive together
 
 	// Play the unit voice response
 	pickAndPlayUnitVoiceResponse(TheInGameUI->getAllSelectedDrawables(), GameMessage::MSG_DO_ATTACKMOVETO);